Add Backbone.Router and React, start rewriting views
After lots of research and discussion with @stephencelis, we decided
to go ahead and try Facebook/React for views. We still need a way
to organize and route requests, so Backbone.Router works for now.

It's been tricky wrapping my mind around how to take full advantage
of the declarative nature of React, but I think I'm getting that hang
of it. I still want to pull a lot of the logic from what's in the
router now into the views, but it's a start.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>