Changes to handle changes in template listed order
In the Makefile we use file to list the jsx files and there is no way to
easily make it sorted by depth. Either we need to do something like what
is in this commit, or maintain the order. For now, I am just going
forward with what I believe is the easiest. I am always open to other
ideas.

Signed-off-by: Tyler Smalley <tylersmalley@me.com>